Switch over all other headers to the new way (minus Timeline)
